✨ Add approve intent
- Add revoke example
