--- src/shared/approve/index.ts
@@ -1,6 +1,0 @@
-import { BaseTool, createToolCollection } from "../client";
-import { intentApproveTool } from "./intents";
-
-export function approveTools(): BaseTool[] {
-  return createToolCollection([intentApproveTool]);
-}
