+++ src/shared/erc20/constants.ts
@@ -1,0 +1,19 @@
+import {
+  arbitrum,
+  base,
+  mainnet,
+  mode,
+  optimism,
+  polygon,
+  sepolia,
+} from "viem/chains";
+
+export const erc20Chains = [
+  mainnet,
+  arbitrum,
+  base,
+  optimism,
+  polygon,
+  mode,
+  sepolia,
+];
